Change quorum threshold type to uint
The type was already being converted to uint
when being passed to SCP, there's no need to
keep it architecture-dependent.

This also allows making the conversion
routines nothrow

+++ source/agora/consensus/Quorum.d
@@ -100,7 +100,7 @@ public QuorumConfig buildQuorumConfig ( const ref PublicKey key,
     }
 
     quorum.nodes.sort;
-    quorum.threshold = max(1, cast(size_t)ceil(THRESHOLD * quorum.nodes.length));
+    quorum.threshold = max(1, cast(uint)ceil(THRESHOLD * quorum.nodes.length));
     return quorum;
 }
